requirements.txt
asgiref==3.7.2
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.4
Django==5.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
flake8==7.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
uritemplate==4.1.1
